chore(migration) refactor sigverify
