+++ app/ante/handler_options_test.go
@@ -107,7 +107,7 @@ func (suite *AnteTestSuite) TestValidateHandlerOptions() {
 				StakingKeeper:   suite.app.StakingKeeper,
 				FeeMarketKeeper: suite.app.FeeMarketKeeper,
 				EvmKeeper:       suite.app.EvmKeeper,
-				SigGasConsumer:  app.SigVerificationGasConsumer,
+				SigGasConsumer:  ante.SigVerificationGasConsumer,
 				SignModeHandler: nil,
 			},
 			false,
@@ -125,7 +125,7 @@ func (suite *AnteTestSuite) TestValidateHandlerOptions() {
 				IBCKeeper:              suite.app.IBCKeeper,
 				FeeMarketKeeper:        suite.app.FeeMarketKeeper,
 				SignModeHandler:        encoding.MakeConfig(app.ModuleBasics).TxConfig.SignModeHandler(),
-				SigGasConsumer:         app.SigVerificationGasConsumer,
+				SigGasConsumer:         ante.SigVerificationGasConsumer,
 				MaxTxGasWanted:         40000000,
 				TxFeeChecker:           ethante.NewDynamicFeeChecker(suite.app.EvmKeeper),
 			},
