+++ app/ante/sigverify_test.go
@@ -1,4 +1,4 @@
-package app
+package ante_test
 
 import (
 	"fmt"
@@ -20,6 +20,8 @@ import (
 
 	"github.com/evmos/ethermint/crypto/ethsecp256k1"
 	"github.com/evmos/ethermint/encoding"
+	"github.com/evmos/evmos/v11/app"
+	"github.com/evmos/evmos/v11/app/ante"
 )
 
 func generatePubKeysAndSignatures(n int, msg []byte, _ bool) (pubkeys []cryptotypes.PubKey, signatures [][]byte) {
@@ -41,7 +43,7 @@ func expectedGasCostByKeys(pubkeys []cryptotypes.PubKey) uint64 {
 		case strings.Contains(pubkeyType, "ed25519"):
 			cost += authtypes.DefaultSigVerifyCostED25519
 		case strings.Contains(pubkeyType, "ethsecp256k1"):
-			cost += secp256k1VerifyCost
+			cost += ante.Secp256k1VerifyCost
 		default:
 			panic("unexpected key type")
 		}
@@ -53,7 +55,7 @@ func TestConsumeSignatureVerificationGas(t *testing.T) {
 	params := authtypes.DefaultParams()
 	msg := []byte{1, 2, 3, 4}
 
-	encodingConfig := encoding.MakeConfig(ModuleBasics)
+	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
 	cdc := encodingConfig.Amino
 
 	p := authtypes.DefaultParams()
@@ -94,7 +96,7 @@ func TestConsumeSignatureVerificationGas(t *testing.T) {
 		{
 			"PubKeyEthsecp256k1",
 			args{sdk.NewInfiniteGasMeter(), nil, ethsecKey.PubKey(), params},
-			secp256k1VerifyCost,
+			ante.Secp256k1VerifyCost,
 			false,
 		},
 		{
@@ -128,7 +130,7 @@ func TestConsumeSignatureVerificationGas(t *testing.T) {
 			Data:     tt.args.sig,
 			Sequence: 0, // Arbitrary account sequence
 		}
-		err := SigVerificationGasConsumer(tt.args.meter, sigV2, tt.args.params)
+		err := ante.SigVerificationGasConsumer(tt.args.meter, sigV2, tt.args.params)
 
 		if tt.shouldErr {
 			require.Error(t, err)
